* Headers/AppKit/NSBrowser.h         * Headers/AppKit/NSFontPanel.h         * Headers/AppKit/NSMatrix.h         * Headers/AppKit/NSTableView.h         * Source/NSBrowser.m         * Source/NSFontPanel.m         * Source/NSMatrix.m: Small fixes to last change.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@36042 72102866-910b-0410-8b05-ffd578937521

+++ Headers/AppKit/NSBrowser.h
@@ -201,7 +201,7 @@ typedef enum _NSBrowserColumnResizingType
 - (void) setTitle: (NSString *)aString
 	 ofColumn: (NSInteger)column;
 - (NSRect) titleFrameOfColumn: (NSInteger)column;
-- (float) titleHeight;
+- (CGFloat) titleHeight;
 - (NSString *) titleOfColumn: (NSInteger)column;
 
 //
